+++ src/Ext/House/Body.h
@@ -18,6 +18,7 @@ public:
 	public:
 		std::map<BuildingTypeExt::ExtData*, int> BuildingCounter;
 		std::map<DWORD, BuildingExt::ExtData*> OwnedLimboDeliveredBuildings;
+		std::vector<TechnoExt::ExtData*> OwnedTimedAutoDeathObjects;
 
 		BuildingClass* Factory_BuildingType;
 		BuildingClass* Factory_InfantryType;
@@ -28,6 +29,7 @@ public:
 		ExtData(HouseClass* OwnerObject) : Extension<HouseClass>(OwnerObject)
 			, BuildingCounter {}
 			, OwnedLimboDeliveredBuildings {}
+			, OwnedTimedAutoDeathObjects {}
 			, Factory_BuildingType { nullptr }
 			, Factory_InfantryType { nullptr }
 			, Factory_VehicleType { nullptr }
@@ -36,6 +38,7 @@ public:
 		{ }
 
 		bool OwnsLimboDeliveredBuilding(BuildingClass const* pBuilding);
+		void UpdateAutoDeathObjectsInLimbo();
 
 		virtual ~ExtData() = default;